+++ test/core/payload/StaticPayloadBuilder.spec.ts
@@ -500,6 +500,60 @@ describe('PayloadBuilder', () => {
                 errorTechnologyType,
             );
         });
+
+        it('should truncate the reason if longer than 1000 chars', () => {
+            // given
+            const reason = 'a'.repeat(1001);
+            const reasonTruncated = 'a'.repeat(1000);
+
+            // given
+            const payload = StaticPayloadBuilder.reportError(
+                name,
+                pAId,
+                sSN,
+                tSSS,
+                reason,
+                ev,
+            );
+
+            // when
+            const { keys, pairs } = parse(payload);
+
+            // then
+            payloadKeysExpect(keys, [
+                PayloadKey.EventType,
+                PayloadKey.ThreadId,
+                PayloadKey.KeyName,
+                PayloadKey.ParentActionId,
+                PayloadKey.StartSequenceNumber,
+                PayloadKey.Time0,
+                PayloadKey.Reason,
+                PayloadKey.ErrorValue,
+                PayloadKey.ErrorTechnologyType,
+            ]);
+
+            payloadExpect(
+                pairs,
+                PayloadKey.EventType,
+                EventType.Error.toString(),
+            );
+            payloadExpect(pairs, PayloadKey.ThreadId, '1');
+            payloadExpect(pairs, PayloadKey.KeyName, name);
+            payloadExpect(pairs, PayloadKey.ParentActionId, pAId.toString());
+            payloadExpect(
+                pairs,
+                PayloadKey.StartSequenceNumber,
+                sSN.toString(),
+            );
+            payloadExpect(pairs, PayloadKey.Time0, tSSS.toString());
+            payloadExpect(pairs, PayloadKey.Reason, reasonTruncated);
+            payloadExpect(pairs, PayloadKey.ErrorValue, ev.toString());
+            payloadExpect(
+                pairs,
+                PayloadKey.ErrorTechnologyType,
+                errorTechnologyType,
+            );
+        });
     });
 
     describe('reportValue', () => {
@@ -762,6 +816,244 @@ describe('PayloadBuilder', () => {
     });
 
     describe('reportCrash', () => {
+        it('should truncate the reason if longer than 1000 chars', () => {
+            // given
+            const reason = 'a'.repeat(1001);
+            const reasonTruncated = 'a'.repeat(1000);
+
+            const stacktrace =
+                'javax.servlet.ServletException: Something bad happened\n' +
+                '    at com.example.myproject.OpenSessionInViewFilter.doFilter(OpenSessionInViewFilter.java:60)\n' +
+                '    at org.mortbay.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1157)\n' +
+                '    at com.example.myproject.ExceptionHandlerFilter.doFilter(ExceptionHandlerFilter.java:28)\n' +
+                '    at org.mortbay.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1157)\n' +
+                '    at com.example.myproject.OutputBufferFilter.doFilter(OutputBufferFilter.java:33)\n' +
+                '    at org.mortbay.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1157)\n' +
+                '    at org.mortbay.jetty.servlet.ServletHandler.handle(ServletHandler.java:388)\n' +
+                '    at org.mortbay.jetty.security.SecurityHandler.handle(SecurityHandler.java:216)\n' +
+                '    at org.mortbay.jetty.servlet.SessionHandler.handle(SessionHandler.java:182)\n' +
+                '    at org.mortbay.jetty.handler.ContextHandler.handle(ContextHandler.java:765)\n' +
+                '    at org.mortbay.jetty.webapp.WebAppContext.handle(WebAppContext.java:418)\n' +
+                '    at org.mortbay.jetty.handler.HandlerWrapper.handle(HandlerWrapper.java:152)\n' +
+                '    at org.mortbay.jetty.Server.handle(Server.java:326)\n' +
+                '    at org.mortbay.jetty.HttpConnection.handleRequest(HttpConnection.java:542)\n' +
+                '    at org.mortbay.jetty.HttpConnection$RequestHandler.content(HttpConnection.java:943)\n' +
+                '    at org.mortbay.jetty.HttpParser.parseNext(HttpParser.java:756)\n' +
+                '    at org.mortbay.jetty.HttpParser.parseAvailable(HttpParser.java:218)\n' +
+                '    at org.mortbay.jetty.HttpConnection.handle(HttpConnection.java:404)\n' +
+                '    at org.mortbay.jetty.bio.SocketConnector$Connection.run(SocketConnector.java:228)\n' +
+                '    at org.mortbay.thread.QueuedThreadPool$PoolThread.run(QueuedThreadPool.java:582)\n';
+
+            const payload = StaticPayloadBuilder.reportCrash(
+                'errorName',
+                reason,
+                stacktrace,
+                6,
+                4000,
+            );
+
+            // when
+            const { keys, pairs } = parse(payload);
+
+            // then
+            payloadKeysExpect(keys, [
+                PayloadKey.EventType,
+                PayloadKey.ThreadId,
+                PayloadKey.KeyName,
+                PayloadKey.ParentActionId,
+                PayloadKey.StartSequenceNumber,
+                PayloadKey.Time0,
+                PayloadKey.Reason,
+                PayloadKey.Stacktrace,
+                PayloadKey.ErrorTechnologyType,
+            ]);
+
+            payloadExpect(
+                pairs,
+                PayloadKey.EventType,
+                EventType.Crash.toString(),
+            );
+            payloadExpect(pairs, PayloadKey.ThreadId, '1');
+            payloadExpect(pairs, PayloadKey.KeyName, 'errorName');
+            payloadExpect(pairs, PayloadKey.ParentActionId, '0');
+            payloadExpect(pairs, PayloadKey.StartSequenceNumber, '6');
+            payloadExpect(pairs, PayloadKey.Time0, '4000');
+            payloadExpect(pairs, PayloadKey.Reason, reasonTruncated);
+            payloadExpect(pairs, PayloadKey.Stacktrace, stacktrace);
+            payloadExpect(
+                pairs,
+                PayloadKey.ErrorTechnologyType,
+                errorTechnologyType,
+            );
+        });
+
+        it('should truncate the stacktrace payload if longer than 128000 chars', () => {
+            // given
+            const stacktrace = 'a'.repeat(128001);
+            const stacktraceTruncated = 'a'.repeat(128000);
+
+            const payload = StaticPayloadBuilder.reportCrash(
+                'errorName',
+                'reason',
+                stacktrace,
+                6,
+                4000,
+            );
+
+            // when
+            const { keys, pairs } = parse(payload);
+
+            // then
+            payloadKeysExpect(keys, [
+                PayloadKey.EventType,
+                PayloadKey.ThreadId,
+                PayloadKey.KeyName,
+                PayloadKey.ParentActionId,
+                PayloadKey.StartSequenceNumber,
+                PayloadKey.Time0,
+                PayloadKey.Reason,
+                PayloadKey.Stacktrace,
+                PayloadKey.ErrorTechnologyType,
+            ]);
+
+            payloadExpect(
+                pairs,
+                PayloadKey.EventType,
+                EventType.Crash.toString(),
+            );
+            payloadExpect(pairs, PayloadKey.ThreadId, '1');
+            payloadExpect(pairs, PayloadKey.KeyName, 'errorName');
+            payloadExpect(pairs, PayloadKey.ParentActionId, '0');
+            payloadExpect(pairs, PayloadKey.StartSequenceNumber, '6');
+            payloadExpect(pairs, PayloadKey.Time0, '4000');
+            payloadExpect(pairs, PayloadKey.Reason, 'reason');
+            payloadExpect(pairs, PayloadKey.Stacktrace, stacktraceTruncated);
+            payloadExpect(
+                pairs,
+                PayloadKey.ErrorTechnologyType,
+                errorTechnologyType,
+            );
+        });
+
+        it('should truncate the stacktrace payload if longer than 128000 chars to the last break', () => {
+            // given
+            const stacktraceTruncated = 'a'.repeat(127900);
+            const stacktrace = stacktraceTruncated + '\n' + 'a'.repeat(1000);
+
+            const payload = StaticPayloadBuilder.reportCrash(
+                'errorName',
+                'reason',
+                stacktrace,
+                6,
+                4000,
+            );
+
+            // when
+            const { keys, pairs } = parse(payload);
+
+            // then
+            payloadKeysExpect(keys, [
+                PayloadKey.EventType,
+                PayloadKey.ThreadId,
+                PayloadKey.KeyName,
+                PayloadKey.ParentActionId,
+                PayloadKey.StartSequenceNumber,
+                PayloadKey.Time0,
+                PayloadKey.Reason,
+                PayloadKey.Stacktrace,
+                PayloadKey.ErrorTechnologyType,
+            ]);
+
+            payloadExpect(
+                pairs,
+                PayloadKey.EventType,
+                EventType.Crash.toString(),
+            );
+            payloadExpect(pairs, PayloadKey.ThreadId, '1');
+            payloadExpect(pairs, PayloadKey.KeyName, 'errorName');
+            payloadExpect(pairs, PayloadKey.ParentActionId, '0');
+            payloadExpect(pairs, PayloadKey.StartSequenceNumber, '6');
+            payloadExpect(pairs, PayloadKey.Time0, '4000');
+            payloadExpect(pairs, PayloadKey.Reason, 'reason');
+            payloadExpect(pairs, PayloadKey.Stacktrace, stacktraceTruncated);
+            payloadExpect(
+                pairs,
+                PayloadKey.ErrorTechnologyType,
+                errorTechnologyType,
+            );
+        });
+
+        it('should truncate the reason if longer than 1000 chars', () => {
+            // given
+            const stacktrace =
+                'javax.servlet.ServletException: Something bad happened\n' +
+                '    at com.example.myproject.OpenSessionInViewFilter.doFilter(OpenSessionInViewFilter.java:60)\n' +
+                '    at org.mortbay.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1157)\n' +
+                '    at com.example.myproject.ExceptionHandlerFilter.doFilter(ExceptionHandlerFilter.java:28)\n' +
+                '    at org.mortbay.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1157)\n' +
+                '    at com.example.myproject.OutputBufferFilter.doFilter(OutputBufferFilter.java:33)\n' +
+                '    at org.mortbay.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1157)\n' +
+                '    at org.mortbay.jetty.servlet.ServletHandler.handle(ServletHandler.java:388)\n' +
+                '    at org.mortbay.jetty.security.SecurityHandler.handle(SecurityHandler.java:216)\n' +
+                '    at org.mortbay.jetty.servlet.SessionHandler.handle(SessionHandler.java:182)\n' +
+                '    at org.mortbay.jetty.handler.ContextHandler.handle(ContextHandler.java:765)\n' +
+                '    at org.mortbay.jetty.webapp.WebAppContext.handle(WebAppContext.java:418)\n' +
+                '    at org.mortbay.jetty.handler.HandlerWrapper.handle(HandlerWrapper.java:152)\n' +
+                '    at org.mortbay.jetty.Server.handle(Server.java:326)\n' +
+                '    at org.mortbay.jetty.HttpConnection.handleRequest(HttpConnection.java:542)\n' +
+                '    at org.mortbay.jetty.HttpConnection$RequestHandler.content(HttpConnection.java:943)\n' +
+                '    at org.mortbay.jetty.HttpParser.parseNext(HttpParser.java:756)\n' +
+                '    at org.mortbay.jetty.HttpParser.parseAvailable(HttpParser.java:218)\n' +
+                '    at org.mortbay.jetty.HttpConnection.handle(HttpConnection.java:404)\n' +
+                '    at org.mortbay.jetty.bio.SocketConnector$Connection.run(SocketConnector.java:228)\n' +
+                '    at org.mortbay.thread.QueuedThreadPool$PoolThread.run(QueuedThreadPool.java:582)\n';
+
+            const reason = 'a'.repeat(1001);
+            const reasonTruncated = 'a'.repeat(1000);
+
+            const payload = StaticPayloadBuilder.reportCrash(
+                'errorName',
+                reason,
+                stacktrace,
+                6,
+                4000,
+            );
+
+            // when
+            const { keys, pairs } = parse(payload);
+
+            // then
+            payloadKeysExpect(keys, [
+                PayloadKey.EventType,
+                PayloadKey.ThreadId,
+                PayloadKey.KeyName,
+                PayloadKey.ParentActionId,
+                PayloadKey.StartSequenceNumber,
+                PayloadKey.Time0,
+                PayloadKey.Reason,
+                PayloadKey.Stacktrace,
+                PayloadKey.ErrorTechnologyType,
+            ]);
+
+            payloadExpect(
+                pairs,
+                PayloadKey.EventType,
+                EventType.Crash.toString(),
+            );
+            payloadExpect(pairs, PayloadKey.ThreadId, '1');
+            payloadExpect(pairs, PayloadKey.KeyName, 'errorName');
+            payloadExpect(pairs, PayloadKey.ParentActionId, '0');
+            payloadExpect(pairs, PayloadKey.StartSequenceNumber, '6');
+            payloadExpect(pairs, PayloadKey.Time0, '4000');
+            payloadExpect(pairs, PayloadKey.Reason, reasonTruncated);
+            payloadExpect(pairs, PayloadKey.Stacktrace, stacktrace);
+            payloadExpect(
+                pairs,
+                PayloadKey.ErrorTechnologyType,
+                errorTechnologyType,
+            );
+        });
+
         it('should build the correct payload', () => {
             // given
             const stacktrace =